Move drop zone out of import csv folder
