+++ src/lib/components/drop-zone/drop-zone.svelte
